Remove redundant props passing by using context directly in StatusBar and Editor.
