+++ new-log-viewer/src/components/Editor/index.tsx
@@ -16,7 +16,11 @@ import {
 import {Nullable} from "../../typings/common";
 import {CONFIG_KEY} from "../../typings/config";
 import {BeginLineNumToLogEventNumMap} from "../../typings/worker";
-import {EDITOR_ACTIONS} from "../../utils/actions";
+import {
+    ACTION_NAME,
+    EDITOR_ACTIONS,
+    handleAction,
+} from "../../utils/actions";
 import {
     CONFIG_DEFAULT,
     getConfig,
@@ -27,22 +31,16 @@ import {
     getMapValueWithNearestLessThanOrEqualKey,
 } from "../../utils/data";
 import MonacoInstance from "./MonacoInstance";
-import {CustomActionCallback} from "./MonacoInstance/typings";
+import {goToPositionAndCenter} from "./MonacoInstance/utils";
 
-
-interface EditorProps {
-    onCustomAction: CustomActionCallback,
-}
 
 /**
  * Renders a read-only editor for viewing logs.
  *
- * @param props
- * @param props.onCustomAction
  * @return
  */
-const Editor = ({onCustomAction}: EditorProps) => {
-    const {logData, beginLineNumToLogEventNum} = useContext(StateContext);
+const Editor = () => {
+    const {beginLineNumToLogEventNum, logData, numEvents} = useContext(StateContext);
     const {logEventNum} = useContext(UrlContext);
 
     const [lineNum, setLineNum] = useState<number>(1);
@@ -51,7 +49,39 @@ const Editor = ({onCustomAction}: EditorProps) => {
     );
     const editorRef = useRef<Nullable<monaco.editor.IStandaloneCodeEditor>>(null);
     const isMouseDownRef = useRef<boolean>(false);
+    const logEventNumRef = useRef<Nullable<number>>(logEventNum);
+    const numEventsRef = useRef<Nullable<number>>(numEvents);
     const pageSizeRef = useRef(getConfig(CONFIG_KEY.PAGE_SIZE));
+
+    const handleEditorCustomAction = useCallback((
+        editor: monaco.editor.IStandaloneCodeEditor,
+        actionName: ACTION_NAME
+    ) => {
+        if (null === logEventNumRef.current || null === numEventsRef.current) {
+            return;
+        }
+        switch (actionName) {
+            case ACTION_NAME.FIRST_PAGE:
+            case ACTION_NAME.PREV_PAGE:
+            case ACTION_NAME.NEXT_PAGE:
+            case ACTION_NAME.LAST_PAGE:
+                handleAction(actionName, logEventNumRef.current, numEventsRef.current);
+                break;
+            case ACTION_NAME.PAGE_TOP:
+                goToPositionAndCenter(editor, {lineNumber: 1, column: 1});
+                break;
+            case ACTION_NAME.PAGE_BOTTOM: {
+                const lineCount = editor.getModel()?.getLineCount();
+                if ("undefined" === typeof lineCount) {
+                    break;
+                }
+                goToPositionAndCenter(editor, {lineNumber: lineCount, column: 1});
+                break;
+            }
+            default:
+                break;
+        }
+    }, []);
 
     /**
      * Sets `editorRef` and configures callbacks for mouse down detection.
@@ -123,9 +153,19 @@ const Editor = ({onCustomAction}: EditorProps) => {
         beginLineNumToLogEventNumRef.current = beginLineNumToLogEventNum;
     }, [beginLineNumToLogEventNum]);
 
+    // Synchronize `logEventNumRef` with `logEventNum`.
+    useEffect(() => {
+        logEventNumRef.current = logEventNum;
+    }, [logEventNum]);
+
+    // Synchronize `numEventsRef` with `numEvents`.
+    useEffect(() => {
+        numEventsRef.current = numEvents;
+    }, [numEvents]);
+
     // On `logEventNum` update, update line number in the editor.
     useEffect(() => {
-        if (null === editorRef.current || true === isMouseDownRef.current) {
+        if (null === editorRef.current || isMouseDownRef.current) {
             // Don't update the line number if the user is actively selecting text.
             return;
         }
@@ -151,7 +191,7 @@ const Editor = ({onCustomAction}: EditorProps) => {
                 lineNum={lineNum}
                 text={logData}
                 onCursorExplicitPosChange={handleCursorExplicitPosChange}
-                onCustomAction={onCustomAction}
+                onCustomAction={handleEditorCustomAction}
                 onMount={handleMount}
                 onTextUpdate={restoreCachedPageSize}/>
         </div>
