+++ new-log-viewer/src/components/Layout.tsx
@@ -1,26 +1,9 @@
-import {
-    useCallback,
-    useContext,
-    useEffect,
-    useRef,
-} from "react";
-
-import * as monaco from "monaco-editor";
-
 import {CssVarsProvider} from "@mui/joy/styles";
 
-import {StateContext} from "../contexts/StateContextProvider";
-import {UrlContext} from "../contexts/UrlContextProvider";
-import {Nullable} from "../typings/common";
 import {CONFIG_KEY} from "../typings/config";
-import {
-    ACTION_NAME,
-    handleAction,
-} from "../utils/actions";
 import {CONFIG_DEFAULT} from "../utils/config";
 import DropFileContainer from "./DropFileContainer";
 import Editor from "./Editor";
-import {goToPositionAndCenter} from "./Editor/MonacoInstance/utils";
 import MenuBar from "./MenuBar";
 import StatusBar from "./StatusBar";
 import monacoTheme from "./theme";
@@ -34,51 +17,6 @@ import "./Layout.css";
  * @return
  */
 const Layout = () => {
-    const {numEvents} = useContext(StateContext);
-    const {logEventNum} = useContext(UrlContext);
-
-    const logEventNumRef = useRef<Nullable<number>>(logEventNum);
-    const numEventsRef = useRef<Nullable<number>>(numEvents);
-
-
-    const handleEditorCustomAction = useCallback((
-        editor: monaco.editor.IStandaloneCodeEditor,
-        actionName: ACTION_NAME
-    ) => {
-        if (null === logEventNumRef.current || null === numEventsRef.current) {
-            return;
-        }
-        switch (actionName) {
-            case ACTION_NAME.FIRST_PAGE:
-            case ACTION_NAME.PREV_PAGE:
-            case ACTION_NAME.NEXT_PAGE:
-            case ACTION_NAME.LAST_PAGE:
-                handleAction(actionName, logEventNumRef.current, numEventsRef.current);
-                break;
-            case ACTION_NAME.PAGE_TOP:
-                goToPositionAndCenter(editor, {lineNumber: 1, column: 1});
-                break;
-            case ACTION_NAME.PAGE_BOTTOM: {
-                const lineCount = editor.getModel()?.getLineCount();
-                if ("undefined" === typeof lineCount) {
-                    break;
-                }
-                goToPositionAndCenter(editor, {lineNumber: lineCount, column: 1});
-                break;
-            }
-            default:
-                break;
-        }
-    }, []);
-
-    useEffect(() => {
-        logEventNumRef.current = logEventNum;
-    }, [logEventNum]);
-
-    useEffect(() => {
-        numEventsRef.current = numEvents;
-    }, [numEvents]);
-
     return (
         <CssVarsProvider
             defaultMode={CONFIG_DEFAULT[CONFIG_KEY.THEME]}
@@ -88,11 +26,9 @@ const Layout = () => {
             <div className={"layout"}>
                 <MenuBar/>
                 <DropFileContainer>
-                    <Editor onCustomAction={handleEditorCustomAction}/>
+                    <Editor/>
                 </DropFileContainer>
-                <StatusBar
-                    logEventNum={logEventNum}
-                    numEvents={numEvents}/>
+                <StatusBar/>
             </div>
         </CssVarsProvider>
     );
